Rename summary app to statement
