+++ statement/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class SummaryConfig(AppConfig):
+class StatementConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "summary"
+    name = "statement"
